requirements.txt
opencv-python
argparse
Pillow